+++ frontend/src/components/Home/Klogo.png
