+++ frontend/src/components/Home/home.css
@@ -1,0 +1,49 @@
+.box {
+    justify-content: center;
+    align-content: center;
+    display: flex;
+}
+
+.nav-buttons {
+    font-family: sofiapro-regular;
+    font-style: normal;
+    font-weight: 500;
+    font-size: 14px;
+    line-height: 120%;
+    text-align: center;
+    color: #333333;
+    border-radius: 100px;
+    margin: 40px 6px 20px 6px;
+    width: 132px;
+    height: 144px;
+    background: #FCFDFF;
+    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+    border-radius: 5px;
+}
+
+.navlogo {
+    margin: 50px;
+}
+
+.forms {
+    border: none;
+    border-bottom: 1px solid #62758B;;
+    font-family: sofiapro-regular;
+    font-style: normal;
+    font-weight: normal;
+    font-size: 16px;
+    line-height: 120%;
+    margin-top: 40px;
+    color: #A4A8AC;
+    width: 300px;
+}
+
+.headline {
+    font-family: sofiapro-regular;
+    font-style: normal;
+    font-weight: bold;
+    font-size: 18px;
+    line-height: 120%;
+    text-align: center;
+    color: #2B8FFF;
+}
